Refactor where to place components
