+++ src/modules/common/components/toggle-theme/toggle-theme.vue
@@ -17,7 +17,7 @@
 </template>
 
 <script setup lang="ts">
-import { Button } from '@/components/ui/button';
+import { Button } from '@/modules/common/components/ui/button';
 import { Icon } from '@iconify/vue';
 import { useColorMode } from '@vueuse/core';
