+++ src/modules/common/components/ui/button/Button.vue
@@ -11,7 +11,7 @@
 
 <script setup lang="ts">
 import type { HTMLAttributes } from 'vue';
-import { cn } from '@/lib/utils';
+import { cn } from '@/modules/common/helpers/utils';
 import { Primitive, type PrimitiveProps } from 'reka-ui';
 import { type ButtonVariants, buttonVariants } from '.';
